Clean-up & improve raycast
