+++ source/tools/common/base.py
@@ -327,12 +327,12 @@ class CarverBase(bpy.types.Operator,
             if tuple(round(v, 4) for v in obj.dimensions) == (0.0, 0.0, 0.0):
                 continue
             if is_linked(context, obj):
-                self.report({'ERROR'}, f"{obj.name} is linked and can not be carved")
+                self.report({'WARNING'}, f"{obj.name} is linked and can not be carved")
                 continue
 
             if self.mode == 'DESTRUCTIVE':
                 if is_instanced_data(obj):
-                    self.report({'ERROR'}, f"Modifiers cannot be applied to {obj.name} because it has instanced object data")
+                    self.report({'WARNING'}, f"Modifiers cannot be applied to {obj.name} because it has instanced object data")
                     continue
 
             selected.append(obj)
@@ -666,7 +666,8 @@ class CarverBase(bpy.types.Operator,
         """Align workplane to the surface normal under the cursor."""
 
         # Cast Ray
-        ray = raycast(context, self.mouse.initial)
+        objects = list(context.scene.objects) if self.align_to_all else self.objects.selected
+        ray = raycast(context, self.mouse.initial, objects)
 
         # Fallback to view alignment if no surface is hit.
         if not ray.hit:
